+++ chat-widget/src/components/livechatwidget/common/endChat.ts
@@ -18,8 +18,8 @@ import { getAuthClientFunction, handleAuthentication } from "./authHelper";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const prepareEndChat = async (props: ILiveChatWidgetProps, chatSDK: any, setAdapter: any, setWebChatStyles: any, dispatch: Dispatch<ILiveChatWidgetAction>, adapter: any, state: ILiveChatWidgetContext) => {
-    const isPostChatEnabled = state.domainStates.liveChatConfig?.LiveWSAndLiveChatEngJoin?.msdyn_postconversationsurveyenable;
-    const postChatSurveyMode = state.domainStates.liveChatConfig?.LiveWSAndLiveChatEngJoin?.msdyn_postconversationsurveymode;
+    const isPostChatEnabled = props.chatConfig?.LiveWSAndLiveChatEngJoin?.msdyn_postconversationsurveyenable ?? state.domainStates.liveChatConfig?.LiveWSAndLiveChatEngJoin?.msdyn_postconversationsurveyenable;
+    const postChatSurveyMode = props.chatConfig?.LiveWSAndLiveChatEngJoin?.msdyn_postconversationsurveymode ?? state.domainStates.liveChatConfig?.LiveWSAndLiveChatEngJoin?.msdyn_postconversationsurveymode;
 
     //Unable to end chat if token has expired
     if (props.getAuthToken) {
@@ -64,13 +64,18 @@ const prepareEndChat = async (props: ILiveChatWidgetProps, chatSDK: any, setAdap
         }
 
         if (postChatSurveyMode === PostChatSurveyMode.Embed) {
-            dispatch({ type: LiveChatWidgetActionType.SET_CONVERSATION_STATE, payload: ConversationState.PostchatLoading });
-            await addDelayInMs(Constants.PostChatLoadingDurationInMs);
+            // Only start embedded Postchat workflow if postchat context is set successfully else close chat
+            if (state.domainStates.postChatContext) {
+                dispatch({ type: LiveChatWidgetActionType.SET_CONVERSATION_STATE, payload: ConversationState.PostchatLoading });
+                await addDelayInMs(Constants.PostChatLoadingDurationInMs);
 
-            const loadPostChatEvent: ICustomEvent = {
-                eventName: BroadcastEvent.LoadPostChatSurvey,
-            };
-            BroadcastService.postMessage(loadPostChatEvent);
+                const loadPostChatEvent: ICustomEvent = {
+                    eventName: BroadcastEvent.LoadPostChatSurvey,
+                };
+                BroadcastService.postMessage(loadPostChatEvent);
+            } else {
+                await endChat(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, true, false, true);
+            }
         } else if (postChatSurveyMode === PostChatSurveyMode.Link) {
             dispatch({ type: LiveChatWidgetActionType.SET_CONVERSATION_STATE, payload: ConversationState.InActive });